feat/#35 editing Security/initdata
