+++ src/main/java/com/beanSpot/WEB3_4_Poten_BE/domain/oauth/OAuth2SuccessHandler.java
@@ -1,6 +1,8 @@
 package com.beanSpot.WEB3_4_Poten_BE.domain.oauth;
 
 import java.io.IOException;
+import java.net.URLEncoder;
+import java.nio.charset.StandardCharsets;
 
 import org.springframework.security.core.Authentication;
 import org.springframework.security.oauth2.client.OAuth2AuthorizedClientService;
@@ -22,7 +24,6 @@ public class OAuth2SuccessHandler extends SimpleUrlAuthenticationSuccessHandler 
 	private final JwtService jwtService;
 	private final OAuth2AuthorizedClientService authorizedClientService;
 
-	@Override
 	public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response,
 		Authentication authentication) throws IOException {
 		SecurityUser securityUser = (SecurityUser) authentication.getPrincipal();
@@ -36,38 +37,18 @@ public class OAuth2SuccessHandler extends SimpleUrlAuthenticationSuccessHandler 
 		log.info("JWT 토큰 생성 완료: accessToken 존재 = {}, refreshToken 존재 = {}",
 			jwtToken != null, refreshToken != null);
 
+		// 토큰 URL 인코딩
+		String encodedToken = URLEncoder.encode(jwtToken, StandardCharsets.UTF_8);
+		String encodedRefreshToken = URLEncoder.encode(refreshToken, StandardCharsets.UTF_8);
 
-
-		/* 쿠키 설정 -> 추후에 인증 방식 설정시 적용
-		ResponseCoo	kie accessTokenCookie = ResponseCookie.from("accessToken", jwtToken)
-			.path("/")
-			.httpOnly(true)
-			.secure(true)
-			.sameSite("None")  // 크로스 사이트 요청에 대한 제한 완화
-			.maxAge(60 * 60)  // 1시간
-			.build();
-
-		ResponseCookie refreshTokenCookie = ResponseCookie.from("refreshToken", refreshToken)
-			.path("/")
-			.httpOnly(true)
-			.secure(true)
-			.sameSite("None")
-			.maxAge(7 * 24 * 60 * 60)  // 7일
-			.build();
-
-		log.info("쿠키 설정: accessToken = {}, refreshToken = {}",
-			accessTokenCookie.toString(), refreshTokenCookie.toString());
-
-		response.addHeader("Set-Cookie", accessTokenCookie.toString());
-		response.addHeader("Set-Cookie", refreshTokenCookie.toString());*/
-
-		// 프론트엔드 리다이렉션 URL (환경에 맞게 조정 필요)
 		// SNS 타입에 따라 다른 콜백 URL로 리다이렉트
 		String snsType = member.getSnsType().toString().toLowerCase();
 		String redirectUrl = String.format(
-			"http://localhost:3000/auth/callback/%s?accessToken=%s&refreshToken=%s",
-			snsType, jwtToken, refreshToken
+			"http://localhost:5173/auth/callback/%s?accessToken=%s&refreshToken=%s",
+			snsType, encodedToken, encodedRefreshToken
 		);
+
+		log.info("리다이렉트 URL: {}", redirectUrl);
 
 		getRedirectStrategy().sendRedirect(request, response, redirectUrl);
 	}
