+++ src/test/java/com/beanSpot/WEB3_4_Poten_BE/domain/admin/AdminControllerTest.java
@@ -23,9 +23,9 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 
-import com.beanSpot.WEB3_4_Poten_BE.domain.admin.controller.AdminController;
+import com.beanSpot.WEB3_4_Poten_BE.domain.admin.controller.AdminMemberController;
 import com.beanSpot.WEB3_4_Poten_BE.domain.admin.dto.AdminLoginDto;
-import com.beanSpot.WEB3_4_Poten_BE.domain.admin.service.AdminService;
+import com.beanSpot.WEB3_4_Poten_BE.domain.admin.service.ReviewAdminService;
 import com.beanSpot.WEB3_4_Poten_BE.domain.application.dto.res.ApplicationRes;
 import com.beanSpot.WEB3_4_Poten_BE.domain.cafe.service.CafeService;
 import com.beanSpot.WEB3_4_Poten_BE.domain.jwt.JwtService;
@@ -46,20 +46,20 @@ public class AdminControllerTest {
 	private PasswordEncoder passwordEncoder;
 
 	@Mock
-	private AdminService adminService;
+	private ReviewAdminService reviewAdminService;
 
 	@Mock
 	private CafeService cafeService;
 
 	@InjectMocks
-	private AdminController adminController;
+	private AdminMemberController adminMemberController;
 
 	private MockMvc mockMvc;
 	private ObjectMapper objectMapper;
 
 	@BeforeEach
 	void setUp() {
-		mockMvc = MockMvcBuilders.standaloneSetup(adminController)
+		mockMvc = MockMvcBuilders.standaloneSetup(adminMemberController)
 			.addFilter((request, response, chain) -> {
 				request.setCharacterEncoding("UTF-8");
 				response.setCharacterEncoding("UTF-8");
@@ -133,7 +133,7 @@ public class AdminControllerTest {
 		ApplicationRes app2 = new ApplicationRes(2L, 102L, "카페2", "서울시 서초구", "010-5678-1234", "PENDING");
 		List<ApplicationRes> applications = Arrays.asList(app1, app2);
 
-		when(adminService.getPendingApplications()).thenReturn(applications);
+		when(reviewAdminService.getPendingApplications()).thenReturn(applications);
 
 		// When & Then
 		mockMvc.perform(get("/admin/applications/pending"))
@@ -154,7 +154,7 @@ public class AdminControllerTest {
 		ApplicationRes approvedApp = new ApplicationRes(
 			applicationId, 101L, "승인된 카페", "서울시 강남구", "010-1234-5678", "APPROVED");
 
-		when(adminService.approveApplication(applicationId)).thenReturn(approvedApp);
+		when(reviewAdminService.approveApplication(applicationId)).thenReturn(approvedApp);
 
 		// When & Then
 		mockMvc.perform(post("/admin/applications/{applicationId}/approve", applicationId))
@@ -163,7 +163,7 @@ public class AdminControllerTest {
 			.andExpect(jsonPath("$.id").value(applicationId))
 			.andExpect(jsonPath("$.status").value("APPROVED"));
 
-		verify(adminService, times(1)).approveApplication(applicationId);
+		verify(reviewAdminService, times(1)).approveApplication(applicationId);
 	}
 
 	@Test
@@ -174,7 +174,7 @@ public class AdminControllerTest {
 		ApplicationRes rejectedApp = new ApplicationRes(
 			applicationId, 101L, "거절된 카페", "서울시 강남구", "010-1234-5678", "REJECTED");
 
-		when(adminService.rejectApplication(applicationId)).thenReturn(rejectedApp);
+		when(reviewAdminService.rejectApplication(applicationId)).thenReturn(rejectedApp);
 
 		// When & Then
 		mockMvc.perform(post("/admin/applications/{applicationId}/reject", applicationId))
@@ -183,7 +183,7 @@ public class AdminControllerTest {
 			.andExpect(jsonPath("$.id").value(applicationId))
 			.andExpect(jsonPath("$.status").value("REJECTED"));
 
-		verify(adminService, times(1)).rejectApplication(applicationId);
+		verify(reviewAdminService, times(1)).rejectApplication(applicationId);
 	}
 
 	@Test
@@ -194,7 +194,7 @@ public class AdminControllerTest {
 		Member member2 = createMember(2L, "owner@test.com", Member.MemberType.OWNER);
 		List<Member> members = Arrays.asList(member1, member2);
 
-		when(adminService.getAllMembers()).thenReturn(members);
+		when(reviewAdminService.getAllMembers()).thenReturn(members);
 
 		// When & Then
 		mockMvc.perform(get("/admin/members"))
@@ -215,7 +215,7 @@ public class AdminControllerTest {
 		Member member2 = createMember(2L, "user2@test.com", Member.MemberType.USER);
 		List<Member> members = Arrays.asList(member1, member2);
 
-		when(adminService.getMembersByType(Member.MemberType.USER)).thenReturn(members);
+		when(reviewAdminService.getMembersByType(Member.MemberType.USER)).thenReturn(members);
 
 		// When & Then
 		mockMvc.perform(get("/admin/members")
@@ -236,7 +236,7 @@ public class AdminControllerTest {
 		Long memberId = 1L;
 		Member member = createMember(memberId, "user@test.com", Member.MemberType.USER);
 
-		when(adminService.getMemberById(memberId)).thenReturn(member);
+		when(reviewAdminService.getMemberById(memberId)).thenReturn(member);
 
 		// When & Then
 		mockMvc.perform(get("/admin/members/{memberId}", memberId))
@@ -245,7 +245,7 @@ public class AdminControllerTest {
 			.andExpect(jsonPath("$.id").value(memberId))
 			.andExpect(jsonPath("$.email").value("user@test.com"));
 
-		verify(adminService, times(1)).getMemberById(memberId);
+		verify(reviewAdminService, times(1)).getMemberById(memberId);
 	}
 
 	@Test
